+++ migrations/001_create_table.sql
@@ -156,7 +156,7 @@ CREATE INDEX IF NOT EXISTS idx_pipelines_progress_pipeline_id ON pipelines_progr
 CREATE TABLE IF NOT EXISTS credit_snowflake_monitoring(
     id SERIAL PRIMARY KEY,
     destination_id INTEGER NOT NULL REFERENCES destinations(id) ON DELETE CASCADE,
-    total_credit INTEGER NOT NULL,
+    total_credit NUMERIC(38, 9) NOT NULL,
     usage_date TIMESTAMP NOT NULL,
     created_at TIMESTAMP DEFAULT NOW(),
     updated_at TIMESTAMP DEFAULT NOW()
